+++ src/loaders/index.ts
@@ -6,3 +6,5 @@ export * from "./MtlLoader.js";
 export * from "./ObjLoader.js";
 export * from "./SkyboxLoader.js";
 export * from "./GltfLoader.js";
+export * from "./BinaryStreamLoader.js";
+export * from "./GeometryWorkerProcessor.js";